+++ app/models/Classroom.php
@@ -127,7 +127,7 @@ class Classroom extends AltActiveRecord
             ];
         }
 
-        return $behaviors;
+        return array_merge($behaviors, parent::behaviors()) ;
     }
 
     /**