+++ app/models/FrequencyAndMeanByDiscipline.php
@@ -16,7 +16,7 @@
  * @property StudentEnrollment $enrollmentFk
  * @property EdcensoDiscipline $disciplineFk
  */
-class FrequencyAndMeanByDiscipline extends CActiveRecord
+class FrequencyAndMeanByDiscipline extends TagModel
 {
 	/**
 	 * @return string the associated database table name
@@ -25,18 +25,7 @@ class FrequencyAndMeanByDiscipline extends CActiveRecord
 	{
 		return 'frequency_and_mean_by_discipline';
 	}
-    public function behaviors()
-    {
-        return [
-            'CTimestampBehavior' => [
-                'class' => 'zii.behaviors.CTimestampBehavior',
-                'createAttribute' => 'created_at',
-                'updateAttribute' => 'updated_at',
-                'setUpdateOnCreate' => true,
-                'timestampExpression' => new CDbExpression('CONVERT_TZ(NOW(), "+00:00", "-03:00")'),
-            ]
-        ];
-    }
+
 
 	/**
 	 * @return array validation rules for model attributes.