+++ app/modules/foods/models/FoodInventorySpent.php
@@ -13,7 +13,7 @@
  * The followings are the available model relations:
  * @property FoodInventory $foodInventoryFk
  */
-class FoodInventorySpent extends CActiveRecord
+class FoodInventorySpent extends TagModel
 {
 	/**
 	 * @return string the associated database table name
@@ -22,18 +22,7 @@ class FoodInventorySpent extends CActiveRecord
 	{
 		return 'food_inventory_spent';
 	}
-    public function behaviors()
-    {
-        return [
-            'CTimestampBehavior' => [
-                'class' => 'zii.behaviors.CTimestampBehavior',
-                'createAttribute' => 'created_at',
-                'updateAttribute' => 'updated_at',
-                'setUpdateOnCreate' => true,
-                'timestampExpression' => new CDbExpression('CONVERT_TZ(NOW(), "+00:00", "-03:00")'),
-            ]
-        ];
-    }
+
 
 	/**
 	 * @return array validation rules for model attributes.