+++ app/modules/foods/models/FoodNoticeVsFoodNoticeItem.php
@@ -12,7 +12,7 @@
  * @property FoodNotice $foodNotice
  * @property FoodNoticeItem $foodNoticeItem
  */
-class FoodNoticeVsFoodNoticeItem extends CActiveRecord
+class FoodNoticeVsFoodNoticeItem extends TagModel
 {
 	/**
 	 * @return string the associated database table name
@@ -21,18 +21,7 @@ class FoodNoticeVsFoodNoticeItem extends CActiveRecord
 	{
 		return 'food_notice_vs_food_notice_item';
 	}
-    public function behaviors()
-    {
-        return [
-            'CTimestampBehavior' => [
-                'class' => 'zii.behaviors.CTimestampBehavior',
-                'createAttribute' => 'created_at',
-                'updateAttribute' => 'updated_at',
-                'setUpdateOnCreate' => true,
-                'timestampExpression' => new CDbExpression('CONVERT_TZ(NOW(), "+00:00", "-03:00")'),
-            ]
-        ];
-    }
+
 
 	/**
 	 * @return array validation rules for model attributes.