+++ app/modules/lunch/models/MenuMeal.php
@@ -12,7 +12,7 @@
  * @property Menu $menuFk
  * @property Meal $mealFk
  */
-class MenuMeal extends CActiveRecord
+class MenuMeal extends TagModel
 {
 	/**
 	 * @return string the associated database table name
@@ -21,18 +21,7 @@ class MenuMeal extends CActiveRecord
 	{
 		return 'lunch_menu_meal';
 	}
-    public function behaviors()
-    {
-        return [
-            'CTimestampBehavior' => [
-                'class' => 'zii.behaviors.CTimestampBehavior',
-                'createAttribute' => 'created_at',
-                'updateAttribute' => 'updated_at',
-                'setUpdateOnCreate' => true,
-                'timestampExpression' => new CDbExpression('CONVERT_TZ(NOW(), "+00:00", "-03:00")'),
-            ]
-        ];
-    }
+
 
 	/**
 	 * @return array validation rules for model attributes.