+++ app/modules/quiz/models/Question.php
@@ -14,7 +14,7 @@
  * @property QuestionOption[] $questionOptions
  * @property Quiz[] $quizs
  */
-class Question extends CActiveRecord
+class Question extends TagModel
 {
 	/**
 	 * @return string the associated database table name
@@ -23,18 +23,7 @@ class Question extends CActiveRecord
 	{
 		return 'question';
 	}
-    public function behaviors()
-    {
-        return [
-            'CTimestampBehavior' => [
-                'class' => 'zii.behaviors.CTimestampBehavior',
-                'createAttribute' => 'created_at',
-                'updateAttribute' => 'updated_at',
-                'setUpdateOnCreate' => true,
-                'timestampExpression' => new CDbExpression('CONVERT_TZ(NOW(), "+00:00", "-03:00")'),
-            ]
-        ];
-    }
+
 
 	/**
 	 * @return array validation rules for model attributes.