+++ app/modules/quiz/models/Quiz.php
@@ -16,7 +16,7 @@
  * @property QuestionGroup[] $questionGroups
  * @property Question[] $questions
  */
-class Quiz extends CActiveRecord
+class Quiz extends TagModel
 {
 	/**
 	 * @return string the associated database table name
@@ -25,18 +25,7 @@ class Quiz extends CActiveRecord
 	{
 		return 'quiz';
 	}
-    public function behaviors()
-    {
-        return [
-            'CTimestampBehavior' => [
-                'class' => 'zii.behaviors.CTimestampBehavior',
-                'createAttribute' => 'created_at',
-                'updateAttribute' => 'updated_at',
-                'setUpdateOnCreate' => true,
-                'timestampExpression' => new CDbExpression('CONVERT_TZ(NOW(), "+00:00", "-03:00")'),
-            ]
-        ];
-    }
+
 
 	/**
 	 * @return array validation rules for model attributes.